requirements.txt
Django==3.0.7
django-createsuperuser
paho_mqtt==1.5.0
django-grappelli==2.14.2
django-smart-selects